requirement.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
celery==4.4.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.8
kombu==4.6.11
Pillow==8.4.0
prompt-toolkit==3.0.20
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
vine==1.3.0
wcwidth==0.2.5
